Big refactoring
Steps are now stored in the store as well, and the Graph and Schedule
operate on References to these. So a graph and a schedule are now only a
collection of references to steps, which points towards a beautiful
variation and extension mechanism, where extensions are considered
as unions of Graphs and variations as diffs (not in set theoretic
sense).